Move TimeoutHelper to new file.
